refactor: split large files into focused modules (ADR-065)
- repl.cpp 1280→302: extracted repl_chat, repl_search, repl_annotations, repl_model
- main.cpp 501→337: extracted sync_annotations
- tui.h 821→70: extracted markdown, table, spinner
- repl_commands.cpp 699→475: extracted repl_model

Also:
- Add 3-tier check system: check-fast, check, check-full
- Add mutation testing via Mull (CI-only, ADR-067)
- Fix clang-tidy braces warnings across all files
- Fix ADR status format for consistency check
- Raise file-size limit to 600 (from 400)
- Fix grep -oP → -oE for macOS compatibility
