+++ .config/mull.yml
@@ -1,0 +1,10 @@
+# mull.yml — Mutation testing configuration
+# @see docs/adr/adr-067-mutation-testing.md
+
+mutators:
+  - cxx_arithmetic
+  - cxx_comparison
+  - cxx_boundary
+  - cxx_logical
+
+timeout: 5000
