+++ .config/versions.env
@@ -42,3 +42,7 @@ GIT_CLIFF_VERSION=2.12.0
 
 # Coverage (lcov/gcov/bc are system packages, not pinned)
 # Runtime (ollama installed via curl installer, not pinned here)
+
+# Mutation testing
+# mull: LLVM-based mutation testing — brew tap on mac, cloudsmith repo on linux
+MULL_VERSION=19
