+++ TECHDEBT.md
@@ -1,36 +1,36 @@
 # Technical Debt Log
 
-Auto-generated by `make todo` on 2026-05-01.
+Auto-generated by `make todo` on 2026-05-02.
 
 ## Code TODOs
 
+- `src/config/config_test.cpp:488:// TODO: test --files with space-separated paths`
+- `src/config/config_test.cpp:489:// TODO: test --files with nonexistent file (should exit with error)`
+- `src/config/config_test.cpp:490:// TODO: test malformed numeric env/CLI values (stoi crash — see TODO.md)`
+- `src/config/config_test.cpp:491:// TODO: test TRACE env var handling (inconsistent between .env and env var)`
+- `src/config/config_test.cpp:492:// TODO: test validate_config rejects invalid port strings like "11434abc"`
 - `src/config/config_test_helper.cpp:51:// clang-tidy:skip-complexity — TODO: refactor (see TODO.md)`
-- `src/config/config_test.cpp:471:// TODO: test --files with space-separated paths`
-- `src/config/config_test.cpp:472:// TODO: test --files with nonexistent file (should exit with error)`
-- `src/config/config_test.cpp:473:// TODO: test malformed numeric env/CLI values (stoi crash — see TODO.md)`
-- `src/config/config_test.cpp:474:// TODO: test TRACE env var handling (inconsistent between .env and env var)`
-- `src/config/config_test.cpp:475:// TODO: test validate_config rejects invalid port strings like "11434abc"`
-- `src/main.cpp:242:// pmccabe:skip-complexity — TODO: refactor main dispatch logic`
-- `src/exec/exec_test.cpp:112:// TODO: test popen failure path (cmd_exec returns "Error: could not execute command")`
-- `src/exec/exec_test.cpp:113:// TODO: test WIFSIGNALED path (command killed by signal)`
+- `src/repl/repl_test.cpp:514:// TODO: test REPL with piped stdin (non-TTY input)`
+- `src/repl/repl_test.cpp:515:// TODO: test /clear resets conversation history`
+- `src/repl/repl_test.cpp:516:// TODO: test /set trace toggles trace mode`
+- `src/repl/repl_test.cpp:517:// TODO: test color_name_to_ansi and ansi_to_name (high complexity, needs refactor)`
+- `src/repl/repl_test.cpp:518:// TODO: test interruptible_chat timeout behavior (needs mock HTTP or thread control)`
+- `src/repl/repl_test.cpp:519:// TODO: test !! command injects output into history`
+- `src/repl/repl_test.cpp:520:// TODO: test ! command does NOT inject output into history`
+- `src/repl/repl_test.cpp:522:// TODO: CSV descriptions in /model (not yet implemented)`
 - `src/json/json_test.cpp:251:// TODO: test json_extract_string with escaped backslash before quote (\\\")`
 - `src/json/json_test.cpp:252:// TODO: test json_extract_object with deeply nested objects`
 - `src/json/json_test.cpp:253:// TODO: test decode_unicode_escape with non-ASCII codepoints (>127)`
 - `src/json/json_test.cpp:254:// TODO: test malformed JSON (unclosed strings, missing braces)`
-- `src/repl/repl_test.cpp:513:// TODO: test REPL with piped stdin (non-TTY input)`
-- `src/repl/repl_test.cpp:514:// TODO: test /clear resets conversation history`
-- `src/repl/repl_test.cpp:515:// TODO: test /set trace toggles trace mode`
-- `src/repl/repl_test.cpp:516:// TODO: test color_name_to_ansi and ansi_to_name (high complexity, needs refactor)`
-- `src/repl/repl_test.cpp:517:// TODO: test interruptible_chat timeout behavior (needs mock HTTP or thread control)`
-- `src/repl/repl_test.cpp:518:// TODO: test !! command injects output into history`
-- `src/repl/repl_test.cpp:519:// TODO: test ! command does NOT inject output into history`
-- `src/repl/repl_test.cpp:521:// TODO: CSV descriptions in /model (not yet implemented)`
+- `src/exec/exec_test.cpp:112:// TODO: test popen failure path (cmd_exec returns "Error: could not execute command")`
+- `src/exec/exec_test.cpp:113:// TODO: test WIFSIGNALED path (command killed by signal)`
 - `src/sync/sync_test.cpp:54:  CHECK (is_read_only("grep -r TODO src/"))`
 - `src/sync/sync_test.cpp:63:  CHECK (is_read_only("cat file.txt | grep TODO"))`
 - `src/sync/sync_test.cpp:88:  CHECK_FALSE (is_read_only("grep TODO | python -c 'import os'"))`
 - `src/logging/logger_test.cpp:173:// TODO: test Logger dev-mode vs installed-mode path detection`
 - `src/logging/logger_test.cpp:174:// TODO: test Logger with unwritable log path (permission denied)`
 - `src/logging/logger_test.cpp:175:// TODO: test Logger escape of all RFC 8259 control characters (\b, \f, \u00xx)`
+- `src/main.cpp:94:// pmccabe:skip-complexity — TODO: refactor main dispatch logic`
 
 **Total code TODOs: 27**
 
@@ -41,6 +41,12 @@ Auto-generated by `make todo` on 2026-05-01.
 - `./docs/task-status.md:65:- [ ] !! High priority task (P1 — do today)`
 - `./docs/task-status.md:66:- [ ] ! Normal priority task (P2 — do this week)`
 - `./docs/task-status.md:67:- [ ]     Low priority task    (P3 — someday)`
+- `./docs/design/tgpt-integration.md:108:- [ ] \`LLMProvider\` interface exists and is used by \`REPL\`.`
+- `./docs/design/tgpt-integration.md:109:- [ ] \`OllamaProvider\` and \`TgptProvider\` implement the interface.`
+- `./docs/design/tgpt-integration.md:110:- [ ] \`/model\` lists models from both providers.`
+- `./docs/design/tgpt-integration.md:111:- [ ] \`/model <N>\` successfully switches the active provider/model.`
+- `./docs/design/tgpt-integration.md:112:- [ ] \`tgpt\` integration is verified via automated tests.`
+- `./docs/design/tgpt-integration.md:113:- [ ] Documentation is updated with new command usage.`
 - `./docs/adr/adr-045-fix-release-pipeline.md:27:- [ ] \`VERSION\` file containing \`0.x.y\` produces tag \`v0.x.y\` and release title \`Release v0.x.y\``
 - `./docs/adr/adr-045-fix-release-pipeline.md:28:- [ ] Release artifacts are named \`llama-cli-<version>-linux-x64.tar.gz\`, \`llama-cli-<version>-linux-arm64.tar.gz\`, \`llama-cli-<version>-macos-arm64.tar.gz\``
 - `./docs/adr/adr-045-fix-release-pipeline.md:29:- [ ] Runner pinned to \`ubuntu-24.04\`, actions pinned to SHAs`
@@ -80,12 +86,6 @@ Auto-generated by `make todo` on 2026-05-01.
 - `./docs/adr/adr-048-quality-framework.md:433:- [ ] Portfolio-level quality dashboard`
 - `./docs/adr/adr-048-quality-framework.md:434:- [ ] Auto-sunset detection for dormant repos`
 - `./docs/prompts/06-todo-scraping.md:8:- The script scans: markdown \`- [ ]\` items and \`TODO|FIXME|HACK|XXX\` in \`src/\` and \`include/\``
-- `./docs/design/tgpt-integration.md:108:- [ ] \`LLMProvider\` interface exists and is used by \`REPL\`.`
-- `./docs/design/tgpt-integration.md:109:- [ ] \`OllamaProvider\` and \`TgptProvider\` implement the interface.`
-- `./docs/design/tgpt-integration.md:110:- [ ] \`/model\` lists models from both providers.`
-- `./docs/design/tgpt-integration.md:111:- [ ] \`/model <N>\` successfully switches the active provider/model.`
-- `./docs/design/tgpt-integration.md:112:- [ ] \`tgpt\` integration is verified via automated tests.`
-- `./docs/design/tgpt-integration.md:113:- [ ] Documentation is updated with new command usage.`
 - `./TODO.md:10:- [ ] config.cpp: \`std::stoi(c.port)\` accepts \`11434abc\` — verify entire port string is numeric`
 - `./TODO.md:11:- [ ] config.cpp: \`--files "/tmp/my notes.txt"\` splits on whitespace — breaks quoted paths`
 - `./TODO.md:12:- [ ] config.cpp: malformed numeric env/CLI values crash before \`validate_config()\` runs (unguarded \`stoi\`)`
