+++ docs/adr/adr-016-tui-design.md
@@ -1,10 +1,6 @@
 # ADR-016: TUI Design
 
-## Status
-
-Accepted
-
-Accepted
+*Status*: Accepted
 
 ## Date
