+++ docs/adr/adr-022-xref-integrity.md
@@ -1,10 +1,6 @@
 # ADR-022: Cross-Reference Integrity Checks
 
-## Status
-
-Proposed
-
-Proposed
+*Status*: Proposed
 
 ## Context
