+++ docs/adr/adr-057-web-search-integration.md
@@ -1,10 +1,6 @@
 # ADR-057: Web Search Integration via Tool Annotations
 
-## Status
-
-Implemented
-
-Implemented
+*Status*: Implemented
 
 ## Context
