+++ docs/adr/adr-062-prompt-format-and-workflow-engine.md
@@ -1,10 +1,6 @@
 # ADR-062: Context-First Architecture & Workflow Engine
 
-## Status
-
-Proposed
-
-Proposed
+*Status*: Proposed
 
 ## Context
